requirements.txt
apipkg==1.4
coverage==4.2
Django==1.10.1
djangorestframework==3.4.3
execnet==1.4.1
factory-boy==2.7.0
fake-factory==0.7.2
freezegun==0.3.7
hypothesis==3.5.3
pep8==1.7.0
pluggy==0.4.0
py==1.4.31
pyflakes==1.3.0
PyHamcrest==1.9.0
pytest==3.0.3
pytest-cache==1.0
pytest-cov==2.4.0
pytest-django==3.0.0
pytest-flakes==1.0.1
pytest-pep8==1.0.6
pytest-xdist==1.15.0
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
spark-parser==1.4.0
tox==2.4.0
uncompyle6==2.9.1
virtualenv==15.0.3
xdis==3.0.2
